refactored ARState
